+++ src/main/java/org/example/JobSearch/service/ResumeService.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 public interface ResumeService {
     void createResume(ResumeDTO resumeDto);
-    void updateResume(Integer resumeId,ResumeDTO resumeDto);
-    void deleteResume(Integer resumeId);
+    void updateResume(Long resumeId,ResumeDTO resumeDto);
+    void deleteResume(Long resumeId);
 
     List<ResumeDTO> getAllResumes();
 
-    List<ResumeDTO> getResumesByCategory(Integer categoryId);
+    List<ResumeDTO> getResumesByCategory(Long categoryId);
 }
